Part 2: Tidy up wording around request object
request object URLs /and/ URNs should be cryptographicly random.

It also seems unsuitable to have a 'shall' clause within the
introduction, so the requirement is moved to the 'response' section.

I also took this opportunity to number the clauses as per other
sections, so they're easier to refer to in external documents etc.

closes #123