Part 1 & 2: Cleanup TLS requirements
Part 1 was still referring to RFC7525, whereas Part 2 had been updated to
refer to BCP195 and also to add extra requirements.

This change moves most of the requirements from Part 2 to Part 1, apart
from 1 clause that I believe is specific to part 2, and updates part 2
to state that the requirements from part 1 apply.

For some reason the part 1 clauses 6.2.1.1 & 6.2.2.1 both overlapped
with the TLS considerations in clause 7.1.

Exceptionally, because the version of part 1 with these requirements
numbered has not been released yet, I believe it is okay to simply
delete these requirements (causing the later clauses in this section
to be renumbered).