Part 2: Require parameters to be inside signed request object
fixes #128 (or at least fixes the only part it seems we can currently
fix without violating the underlying specifications)

This should help mitigate any attacks that require changing the
redirect_uri etc.

The redirect_uri appears to still be required as a parameter outside
the request object for compliance with the current underlying OAuth RFC
(although my personal opinion is that this is not clearly stated in
the specifications). response_type, scope and client_id must be
included as parameters outside the request object.

The OIDC Core spec is already clear on behaviour in the case where
these fields are present in both location:

"When the request parameter is used, the OpenID Connect request
parameter values contained in the JWT supersede those passed using the
OAuth 2.0 request syntax."