Bring access token requirements inline with RFC6749 section 10.10
Our previous minimum on the length of the token appears to be lower
than RFC6749 calls for.

To address this we take two actions:

1. Change from 128 bit token to token with 128 bits of entropy

2. We take the 'should' clause from RFC6749, turn it into a shall
(which seems reasonable given FAPI needs to be hitting the high end
of any security requirement, and add a practical example to avoid
misinterpretation of the requirement.

Based on the assumption that a token can only have a maximum of 6 bits
of entropy for each 8 bit byte (which mainly comes from the requirement
that tokens can only use some subset of ASCII) this implies the token
needs to be at least 22 characters long.

It is left as an exercise for the implementor how to fully achieve the
guessability restriction in RFC6749, for example through the use of
rate limiting or IP blacklisting for repeated attempst to use bad
access tokens.

fixes #132