Part 1: Avoid renumbering of clauses since previous drafts
Commit 4b251dd2f7, when adding the clause about duplicate client ids,
renumbered ~10 spec clauses in Part 1 section 5.2.2

Avoid that by moving the new clause to the end of the requirements,
and adding a new section (and hence new numbering) for the clauses
surrounding returning authenticated user's identifier.

This does introduce a renumbering of clauses, which is
regrettable, but probably the lesser of two evils.