FAPI 1 - Read Profile
6.2 Read-only access provisions
6.2.1 Protected resources provisions
The resource server with the FAPI endpoints:

8. shall encode the response in UTF-8;

The Open Banking Financial APIs provided data in formats that are not necessarily TEXT. i.e CSV's PDF's.

Clause 9. shall send the Content-type HTTP header Content-Type: application/json; charset=UTF-8 **if applicable;** makes it clear that the Content-type response should be UTF-8 only if it makes sense.

I suggest adding the "if applicable;" to the end of clause 8 as well. A more explicit alternative could be "shall encode the response in UTF-8 if the resource is text";