Merged in josephheenan/fapi/restrict-aud (pull request #95)
FAPI-RW: Require the 'aud' claim in the request object

Currently the request object an RP generates for one OP may well be
accepted by another OP. This seems undesireable and hence we
strengthen the 'should's found in
https://openid.net/specs/openid-connect-core-1_0.html#RequestObject to
MUSTs, as discussed is #190.

It is left unspecified what happens if a client sends a request object
where the aud is an array, particularly in the case where it contains
audiences other than the AS's issuer.

closes #190
closes #214

Approved-by: Brian Campbell <bcampbell@pingidentity.com>
Approved-by: Dave Tonge <dave.tonge@bluespeckfinancial.co.uk>