FAPI-RW: Make MTLS the only holder of key mechanism
At the present time, OAuth Token Binding has not been widely deployed,
so in the interests of interoperability it is better to focus
implementors attention on MTLS certificate bound access tokens at the
current time - MTLS has been used in many FAPI deployments.

This situation is likely to change again in later drafts if other methods
of sending constraining tokens are developed.

closes #225