Merged in auth-endpoint-id-token (pull request #180)
Add note that AS may return fewer claims in auth endpoint id token

* Add note that AS may return fewer claims in auth endpoint id token

This is just to reenforce that there is no need to return privacy
sensitive claims in the id_token from the authorization endpoint, as
the sensitive claims can be included in the id_token returned from the
token endpoint.

This is already explicitly allowed by the OpenID Connect spec:

https://openid.net/specs/openid-connect-core-1_0.html#HybridIDToken2

Also convert an existing note into a clause, as it contains a 'should'.

closes #264

Approved-by: Dave Tonge